Proc attribute integer literals...
Hello! I ran into this issue when following along with the tutorials.

Firstly they are amazing and truly helpful in the GL space. I've really loved them. So thank you.

My guess is that this is due to a change in rustc since the tutorials were posted. Basically this is a simple change to support integer literals on top of string literals in the procedural macro attributes.